+++ graft/coreth/RELEASES.md
@@ -2,6 +2,9 @@
 
 ## [v0.12.5](https://github.com/ava-labs/coreth/releases/tag/v0.12.5)
 
+- Migrate geth changes from v1.11.4 through v1.12.0
+- Bump AvalancheGo dependency to v1.10.6-rc.4 to use sized LRU for Chain State Cache
+
 ## [v0.12.4](https://github.com/ava-labs/coreth/releases/tag/v0.12.4)
 
 - Fix API handler crash for `lookupState` in `prestate` tracer